trans(sr) - Getting Started minus Help System
switched to poedit, still have to figure out
spell check integration (i guess an option in
the pro version of poedit?)

apologies, in Serbian "Netherlands" is "Holland",
there is no word for "Netherlands", but since its
an exonym, i guess it's fine

NOTE some people will argue that Serbian must use
cyrillic script, but that's BS, both latin and cyrillic
are taught in schools, I used latin so as to make the
translation useful to speakers of other languages not
familiar with cyrillic.

I begin to notice ambiguities in the original strings.
Luckily I'm a cross-platform kind of nerd, so I can
regurgitate all the irregularities; Some parts of the
manual are spiking in technicality, so I tried to simplify.

I always have the guidelines at hand, as kind of a checklist
of dos and donts and doh's...

see ya, that was a fine day for my keyboard. hailing
from FreeBSD on VirtualBox,
Canadair (aka ______)